backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
sqlalchemy[asyncio]==2.0.36
asyncpg==0.30.0
pydantic-settings==2.7.1
python-multipart==0.0.20
alembic==1.14.1
aiosmtplib==3.0.2
slowapi==0.1.9
